+++ lib/ember_cli/trailing_slash_constraint.rb
@@ -1,0 +1,7 @@
+module EmberCli
+  class TrailingSlashConstraint
+    def matches?(request)
+      !request.original_fullpath.to_s.ends_with?("/")
+    end
+  end
+end